+++ .github/workflows/dvp_fabric_fabtoken.yml
@@ -1,4 +1,4 @@
-name: Token Basics with DLOG Driver (Orion)
+name: Token DVP with FabToken Driver (Fabric)
 
 on:
   push:
@@ -31,4 +31,4 @@ jobs:
       run: make checks
 
     - name: Integration Tests
-      run: make integration-tests-tcc-dlog-orion
+      run: make integration-tests-tcc-dvp-fabtoken
