+++ .github/workflows/fungible_fabric_fabtoken.yml
@@ -1,4 +1,4 @@
-name: NFT with FabToken Driver
+name: Fungible Tokens with FabToken Driver (Fabric)
 
 on:
   push:
@@ -31,4 +31,4 @@ jobs:
       run: make checks
 
     - name: Integration Tests
-      run: make integration-tests-tcc-nft-fabtoken
+      run: make integration-tests-tcc-fabtoken-fabric
