+++ integration/ports.go
@@ -32,6 +32,8 @@ const (
 	FabTokenDVP
 	ZKATDLogNFT
 	FabTokenNFT
+	OrionZKATDLogNFT
+	OrionFabTokenNFT
 )
 
 // StartPortForNode On linux, the default ephemeral port range is 32768-60999 and can be
