+++ integration/token/tcc/fungible/dlog/dlog_test.go
@@ -28,7 +28,7 @@ var _ = Describe("EndToEnd", func() {
 	Describe("Fungible", func() {
 		BeforeEach(func() {
 			var err error
-			network, err = integration.New(StartPortDlog(), "", fungible.Topology("dlog")...)
+			network, err = integration.New(StartPortDlog(), "", fungible.Topology("fabric", "dlog")...)
 			Expect(err).NotTo(HaveOccurred())
 			network.RegisterPlatformFactory(token.NewPlatformFactory())
 			network.Generate()
