+++ integration/token/tcc/fungible/ofabtoken/fabtoken_test.go
@@ -8,11 +8,11 @@ package fabtoken
 
 import (
 	"github.com/hyperledger-labs/fabric-smart-client/integration"
+	"github.com/hyperledger-labs/fabric-token-sdk/integration/token/tcc/fungible"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 
 	"github.com/hyperledger-labs/fabric-token-sdk/integration/nwo/token"
-	"github.com/hyperledger-labs/fabric-token-sdk/integration/token/tcc/basic"
 )
 
 var _ = Describe("Orion EndToEnd", func() {
@@ -28,7 +28,7 @@ var _ = Describe("Orion EndToEnd", func() {
 	Describe("Orion FabToken", func() {
 		BeforeEach(func() {
 			var err error
-			network, err = integration.New(StartPortDlog(), "", basic.Topology("orion", "fabtoken")...)
+			network, err = integration.New(StartPortDlog(), "", fungible.Topology("orion", "fabtoken")...)
 			Expect(err).NotTo(HaveOccurred())
 			network.RegisterPlatformFactory(token.NewPlatformFactory())
 			network.Generate()
@@ -36,7 +36,7 @@ var _ = Describe("Orion EndToEnd", func() {
 		})
 
 		It("succeeded", func() {
-			basic.TestAll(network)
+			fungible.TestAll(network)
 		})
 	})
