+++ integration/token/tcc/nft/dlog/dlog_test.go
@@ -26,7 +26,7 @@ var _ = Describe("EndToEnd", func() {
 	Describe("NFT", func() {
 		BeforeEach(func() {
 			var err error
-			network, err = integration.New(StartPortDlog(), "", nft.Topology("fabric", "dlog")...)
+			network, err = integration.New(StartPortDlog(), "", nft.Topology("dlog")...)
 			Expect(err).NotTo(HaveOccurred())
 			network.RegisterPlatformFactory(token.NewPlatformFactory())
 			network.Generate()
