+++ integration/token/tcc/nft/fabtoken/fabtoken_test.go
@@ -26,7 +26,7 @@ var _ = Describe("EndToEnd", func() {
 	Describe("NFT", func() {
 		BeforeEach(func() {
 			var err error
-			network, err = integration.New(StartPortDlog(), "", nft.Topology("fabric", "fabtoken")...)
+			network, err = integration.New(StartPortDlog(), "", nft.Topology("fabtoken")...)
 			Expect(err).NotTo(HaveOccurred())
 			network.RegisterPlatformFactory(token.NewPlatformFactory())
 			network.Generate()
